+++ src/package.lisp
@@ -42,7 +42,6 @@
    ;; types and stuff
    #:*default-assemblies*
    #:resolve-type
-   #:current-app-domain
    #:load-assembly
    #:load-assembly-from
    #:import-assembly