+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.7.1 Fix compatibility with previous versions
+  * ```GET-LOADED-ASSEMBLIES`` now utilizes interop library method and does not accept any arguments
+  * Removed ```CURRENT-APP-DOMAIN``` function from the list of exported functions
+
 # 0.7.0 .Net Core 3.0 compatibility
   * Added few fixes for .Net Core 3.0 compatibility
   * ```GET-LOADED-ASSEMBLIES``` now accepts optional ```AppDomain``` argument